+++ app/Http/Controllers/ServiceController.php
@@ -26,10 +26,24 @@ class ServiceController extends Controller
      */
     public function index()
     {
-        $services = Service::all();
+        $this->authorize('viewAny', Service::class);
+
+        $services = Service::orderBy('order', 'ASC')->get();
         return view('admin.service.index')->with([
             'services' => $services
         ]);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        $this->authorize('create', Service::class);
+
+        return view('admin.service.create');
     }
 
     /**
@@ -40,23 +54,52 @@ class ServiceController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Service::class);
+
         $request->validate([
             'title' => ['required', 'max:255'],
             'description' => ['required', 'max:255'],
-            'logo_image_url' => ['required', 'max:512'],
+            'logo_image_url' => ['required', 'mimes:png,jpg,jpeg', 'max:512'],
             'order' => ['required', 'min:1', 'max:10']
         ]);
+
+        $logoImage = $request->file('logo_image_url');
+
+        if ($logoImage && $logoImage->isValid()) {
+            $logoImageName = time() . '-' . $logoImage->getClientOriginalName();
+            $logoImagePath = $logoImage->storeAs('public/uploads', $this->pathNormalizer->normalizePath($logoImageName));
+            if (!$logoImagePath) {
+                $this->banner('Logo image upload failed!', 'error');
+                return redirect()->route('service.index');
+            }
+        }
 
         Service::create([
             'user_id' => Auth()->user()->id,
             'title' => $request->title,
-            'logo_image_url' => $request->logo_image_url,
+            'logo_image_url' => $logoImagePath,
             'description' => $request->description,
             'order' => intval($request->order),
         ]);
 
         $this->banner('Successfully created the service!');
         return redirect()->route('service.index');
+    }
+
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Service  $service
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Service $service)
+    {
+        $this->authorize('view', Service::class);
+
+        return view('admin.service.show')->with([
+            'service' => $service,
+        ]);
     }
 
     /**
@@ -68,19 +111,40 @@ class ServiceController extends Controller
      */
     public function update(Request $request, Service $service)
     {
+        $this->authorize('update', Service::class);
+
         $request->validate([
             'title' => ['required', 'max:255'],
             'description' => ['required', 'max:255'],
-            'logo_image_url' => ['required', 'max:512'],
+            'logo_image_url' => ['nullable', 'mimes:png,jpg,jpeg', 'max:512'],
             'order' => ['required', 'min:1', 'max:10']
         ]);
 
-        $service->update(['user_id' => Auth()->user()->id,
+        $data = [
+            'user_id' => Auth()->user()->id,
             'title' => $request->title,
-            'logo_image_url' => $request->logo_image_url,
             'description' => $request->description,
             'order' => intval($request->order),
-        ]);
+        ];
+
+        $logoImage = $request->file('logo_image_url');
+
+        if ($logoImage && $logoImage->isValid()) {
+            $logoImageName = time() . '-' . $logoImage->getClientOriginalName();
+            $logoImagePath = $logoImage->storeAs('public/uploads', $this->pathNormalizer->normalizePath($logoImageName));
+            if (!$logoImagePath) {
+                $this->banner('Logo image upload failed!', 'error');
+                return redirect()->route('service.index');
+            }
+
+            if ($service->logo_image_url) {
+                Storage::delete($service->logo_image_url);
+            }
+
+            $data['logo_image_url'] = $logoImagePath;
+        }
+
+        $service->update($data);
 
         $this->banner('Successfully saved the service!');
         return redirect()->route('service.index');
@@ -94,7 +158,13 @@ class ServiceController extends Controller
      */
     public function destroy(Service $service)
     {
+        $this->authorize('delete', Service::class);
+
         $oldTitle = htmlentities($service->title);
+        if ($service->logo_image_url) {
+            Storage::delete($service->logo_image_url);
+        }
+
         $service->deleteOrFail();
 
         $this->banner('Successfully deleted the service "' . htmlentities($oldTitle) . '"!');
